Added the second SQL and made sure all tests were passing.

+++ server/src/main/java/Services/LogoutService.java
@@ -1,6 +1,7 @@
 package Services;
 
 import DataAccess.AuthDAO;
+import dataAccess.DataAccessException;
 import model.AuthData;
 
 public class LogoutService {
@@ -15,6 +16,9 @@ public class LogoutService {
         AuthData token;
         try {
             token = this.authDAO.getAuth(authToken);
+            if (token == null) {
+                throw new DataAccessException("This token doesn't exist.");
+            }
             this.authDAO.deleteAuth(token.authToken());
         } catch (dataAccess.DataAccessException e) {
             throw new dataAccess.DataAccessException("this token probably doesn't exist.");
